moved input validation into TextArea dir
